refactor(plugins): Use different max gas defaults for execute/fill
